requirements.txt
# Core
Django==2.0.3

# Rest apis
djangorestframework==3.7.7
djangorestframework-simplejwt==3.2.3
djangorestframework-jwt==1.11.0
django-rest-swagger==2.1.2
django-filter==1.1.0
django-cors-headers==2.2.0